+++ ios/docs/html/protocol_c_m_serializable-p.js
@@ -1,4 +1,4 @@
-var protocol_c_m_serializable-p =
+var protocol_c_m_serializable_p =
 [
     [ "className", "protocol_c_m_serializable-p.html#a1fba26b689afa61b794bb4d333b7bc1b", null ],
     [ "objectId", "protocol_c_m_serializable-p.html#a030e4239b227f0097e86688b5a4f4947", null ]