+++ ios/docs/html/search/all_75.js
@@ -1,5 +1,6 @@
 var searchData=
 [
+  ['uiviewcontroller_28modal_29',['UIViewController(Modal)',['../category_u_i_view_controller_07_modal_08.html',1,'']]],
   ['units',['units',['../interface_c_m_distance.html#a6a06adc8cf522bba5d0eda89cbbd9a3e',1,'CMDistance']]],
   ['updateacl_3auser_3asuccesshandler_3aerrorhandler_3a',['updateACL:user:successHandler:errorHandler:',['../interface_c_m_web_service.html#ac71c26f9ba0fd74e482837f1620fe566',1,'CMWebService']]],
   ['updatevaluesfromdictionary_3aserversidefunction_3auser_3aextraparameters_3asuccesshandler_3aerrorhandler_3a',['updateValuesFromDictionary:serverSideFunction:user:extraParameters:successHandler:errorHandler:',['../interface_c_m_web_service.html#a64e982500022479a7305803515607cbe',1,'CMWebService']]],