+++ ios/ios/src/Users/CMUser.m
@@ -67,21 +67,8 @@ static CMWebService *webService;
     }
 }
 
-- (id)init
-{
-    if (self = [super init]) {
-        self.token = nil;
-        self.userId = nil;
-        self.password = nil;
-        self.services = nil;
-        objectId = @"";
-        if (!webService) {
-            webService = [[CMWebService alloc] init];
-        }
-        isDirty = NO;
-        [self registerAllPropertiesForKVO];
-    }
-    return self;
+- (id)init {
+    return [self initWithUserId:nil andPassword:nil];
 }
 
 - (id)initWithUserId:(NSString *)theUserId andPassword:(NSString *)thePassword {
@@ -89,6 +76,7 @@ static CMWebService *webService;
         self.token = nil;
         self.userId = theUserId;
         self.password = thePassword;
+        self.services = nil;
         objectId = @"";
         if (!webService) {
             webService = [[CMWebService alloc] init];
@@ -108,6 +96,7 @@ static CMWebService *webService;
         token = [coder decodeObjectForKey:@"token"];
         tokenExpiration = [coder decodeObjectForKey:@"tokenExpiration"];
         userId = [coder decodeObjectForKey:@"userId"];
+        services = [coder decodeObjectForKey:@"services"];
         if (!webService) {
             webService = [[CMWebService alloc] init];
         }
@@ -168,6 +157,7 @@ static CMWebService *webService;
     if (self.userId) {
         [coder encodeObject:self.userId forKey:@"userId"];
     }
+    [coder encodeObject:self.services forKey:@"services"];
 }
 
 #pragma mark - Comparison
@@ -347,10 +337,11 @@ static CMWebService *webService;
 #pragma mark - Social login with Singly
 
 // This code is very similar to login above, perhaps we can refactor.
--(void)loginWithSocial:(NSString *)service  andViewController:(UIViewController *)viewController callback:(CMUserOperationCallback)callback {
-    [webService loginWithSocial:self withService:service andViewController:viewController callback:^(CMUserAccountResult result, NSDictionary *responseBody) {
+- (void)loginWithSocialNetwork:(NSString *)service viewController:(UIViewController *)viewController params:(NSDictionary *)params callback:(CMUserOperationCallback)callback {
+    
+    [webService loginWithSocial:self withService:service viewController:viewController params:params callback:^(CMUserAccountResult result, NSDictionary *responseBody) {
         NSArray *messages = [NSArray array];
-
+        
         if (result == CMUserAccountLoginSucceeded) {
             self.token = [responseBody objectForKey:@"session_token"];
             
@@ -360,6 +351,8 @@ static CMWebService *webService;
             self.tokenExpiration = [df dateFromString:[responseBody objectForKey:@"expires"]];
             NSDictionary *userProfile = [responseBody objectForKey:@"profile"];
             objectId = [userProfile objectForKey:CMInternalObjectIdKey];
+            
+            self.services = [[responseBody objectForKey:@"profile"] objectForKey:@"__services__"];
 
             if (!self.isDirty) {
                 // Only bring the changes from the server into the object state if there weren't local modifications.