+++ ios/ios/src/Web Services/CMSocialLoginViewController.h
@@ -19,12 +19,13 @@
 
 @property (nonatomic,strong) id<CMSocialLoginViewControllerDelegate> delegate;
 
-@property (strong, atomic) NSString *targetService;
-@property (strong, atomic) NSString *appID;
-@property (strong, atomic) NSString *apiKey;
-@property (strong, atomic) NSString *challenge;
-@property (strong, atomic) NSString *session_token;
-@property (strong, atomic) CMUser *user;
+@property (strong, nonatomic) NSString *targetService;
+@property (strong, nonatomic) NSString *appID;
+@property (strong, nonatomic) NSString *apiKey;
+@property (strong, nonatomic) NSString *challenge;
+@property (strong, nonatomic) NSString *session_token;
+@property (strong, nonatomic) NSDictionary *params;
+@property (strong, nonatomic) CMUser *user;
 
 
 
@@ -32,9 +33,13 @@
  *
  * Initialize with a service identifier
  *
- * @param service The name of the service that we are logging into.
+ * @param service The name of the service that we are logging into
+ * @param appID The appID gotten from the dashboard
+ * @param apiKey The APIKey from the dashboard for your application
+ * @param user Can be nil, the user you want to link accounts with. If this parameter is nil, we will not link the accounts. If you pass in the user, we will attempt to link the accounts.
+ * @param params Any extra parameters you want passed in to the authentication request. This dictionary is parsed where each key value pair becomes "&key=value". We do not encode the URL after this, so any encoding will need to be done by the creator. This is a good place to put scope, for example: @{@"scope" : @"gist,repo"}
  */
-- (id)initForService:(NSString *)service withAppID:(NSString *)appID andApiKey:(NSString *)apiKey user:(CMUser *)user;
+- (id)initForService:(NSString *)service appID:(NSString *)appID apiKey:(NSString *)apiKey user:(CMUser *)user params:(NSDictionary *)params;
 
 @end